app/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
python-dotenv>=1.0.0
asyncpg>=0.26.0
jinja2>=3.1.2
slowapi>=0.1.9